+++ src/Turnierplan.Core/Planning/InvitationLinkEntry.cs
@@ -2,16 +2,16 @@ using Turnierplan.Core.SeedWork;
 
 namespace Turnierplan.Core.Planning;
 
-public sealed class InvitationLinkEntry : Entity<int>
+public sealed class InvitationLinkEntry : Entity<long>
 {
-    internal InvitationLinkEntry(int id, int? maxTeamsPerRegistration, bool allowNewRegistrations)
+    internal InvitationLinkEntry(long id, int? maxTeamsPerRegistration, bool allowNewRegistrations)
     {
         Id = id;
         MaxTeamsPerRegistration = maxTeamsPerRegistration;
         AllowNewRegistrations = allowNewRegistrations;
     }
 
-    public override int Id { get; protected set; }
+    public override long Id { get; protected set; }
 
     public TournamentClass Class { get; internal set; } = null!;
